apps/python/pyproject.toml
[project]
name = "cc-feishu-python"
version = "0.1.0"
description = "Feishu bot with Claude Code CLI integration"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "lark-oapi>=1.0.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
